Move LEXGINE to GIT
The current version is on early stage of development and
contains selection of low-level object for more convenient
interaction with Direct3D 12 API. All objects are designed
following the open-closed principle, i.e. it is always
possible to interlieve higher-lever calls with the most
basic API interaction. However, consistency of such
interaction is a user's burden